:hammer: Refactor series details view classes directory
