+++ app/src/main/res/layout/fragment_episode_details.xml
@@ -5,7 +5,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".view.SeriesDetailsFragment">
+    tools:context=".view.seriesDetails.SeriesDetailsFragment">
 
     <androidx.appcompat.widget.AppCompatImageView
         android:id="@+id/acivEpisodeDetailBanner"
