+++ app/src/main/res/navigation/nav_graph.xml
@@ -17,7 +17,7 @@
     </fragment>
     <fragment
         android:id="@+id/SeriesDetailsFragment"
-        android:name="com.rummenigged.popcorntime.view.SeriesDetailsFragment"
+        android:name="com.rummenigged.popcorntime.view.seriesDetails.SeriesDetailsFragment"
         android:label="@string/series_details_fragment_label"
         tools:layout="@layout/fragment_series_details">
